refactored :listener_interface
